downloader: fix invalid chars in album/artist name
If the name of an artist, an album
contains the forward slash character, a.k.a. `/`,
ncm will failed to create the directory to save the download file
and will abort its execution with an exception.

For example, the album `雨のち晴れ / What's your name?`
contains `/`, which is forbidden in unix systems.

https://music.163.com/#/album?id=2569036

Commit `cd3d043` replaces some special characters with white space
in track name,
but does not replace special characters in album and artist names.

This commit fix this issue by also replacing special characters
in album and artist names.

Related issue: #7